Remove Point Cloud Queue in imageProjection
I saw there was a queue for incoming point cloud messages. I see no benefit of having it in the code. It also ruines the real-time capabilities of the LIO-SAM, by processing messages that are in the past, leads to unwanted delay. I have checked that IMU and ODOM messages can be linked with "timeScanCur" with no problems.
I propose the removal of this queue.